Split controller.go off of the Configuration reconciler
